refactor: comment 도메인 관련 컴포넌트를 하나의 폴더로 모음
